+++ src/DevilDaggersInfo.Core.Replay/ReplayBinary.cs
@@ -1,5 +1,3 @@
-using DevilDaggersInfo.Core.Replay.Events.Data;
-using DevilDaggersInfo.Core.Replay.Events.Enums;
 using System.Diagnostics.CodeAnalysis;
 
 namespace DevilDaggersInfo.Core.Replay;
@@ -20,30 +18,17 @@ public class ReplayBinary<TReplayBinaryHeader>
 		else
 			compressedDataLength = (int)(contents.Length - br.BaseStream.Position);
 
-		EventsData = ReplayEventsParser.Parse(br.ReadBytes(compressedDataLength));
+		Events = ReplayEventsParser.Parse(br.ReadBytes(compressedDataLength));
 	}
 
-	public ReplayBinary(TReplayBinaryHeader header, byte[] compressedEvents)
+	public ReplayBinary(TReplayBinaryHeader header, IReadOnlyList<ReplayEvent> events)
 	{
 		Header = header;
-		EventsData = ReplayEventsParser.Parse(compressedEvents);
-	}
-
-	public ReplayBinary(TReplayBinaryHeader header, ReplayEventsData eventsData)
-	{
-		Header = header;
-		EventsData = eventsData;
+		Events = events;
 	}
 
 	public TReplayBinaryHeader Header { get; }
-	public ReplayEventsData EventsData { get; }
-
-	public static ReplayBinary<TReplayBinaryHeader> CreateDefault()
-	{
-		return new(
-			header: TReplayBinaryHeader.CreateDefault(),
-			compressedEvents: ReplayEventsCompiler.CompileEvents([new(new InitialInputsEventData(false, false, false, false, JumpType.None, ShootType.None, ShootType.None, 0, 0, 2)), new(new EndEventData())]));
-	}
+	public IReadOnlyList<ReplayEvent> Events { get; }
 
 	public static bool TryParse(byte[] fileContents, [NotNullWhen(true)] out ReplayBinary<TReplayBinaryHeader>? replayBinary)
 	{
@@ -66,7 +51,7 @@ public class ReplayBinary<TReplayBinaryHeader>
 
 		bw.Write(Header.ToBytes());
 
-		byte[] compressedEvents = ReplayEventsCompiler.CompileEvents(EventsData.Events.ToList());
+		byte[] compressedEvents = ReplayEventsCompiler.CompileEvents(Events);
 		bw.Write(compressedEvents.Length);
 		bw.Write(compressedEvents);